[Refactor #37]  코드리뷰에 대한 리펙토링
