+++ app/livekit/get-participant-token/route.ts
@@ -1,9 +1,9 @@
-import { AccessToken } from 'livekit-server-sdk';
-import { NextRequest, NextResponse } from 'next/server';
+import { AccessToken } from "livekit-server-sdk";
+import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(req: NextRequest) {
-  const room = req.nextUrl.searchParams.get('room');
-  const username = req.nextUrl.searchParams.get('username');
+  const room = req.nextUrl.searchParams.get("room");
+  const username = req.nextUrl.searchParams.get("username");
   if (!room) {
     return NextResponse.json({ error: 'Missing "room" query parameter' }, { status: 400 });
   } else if (!username) {
@@ -15,7 +15,7 @@ export async function GET(req: NextRequest) {
   const wsUrl = process.env.NEXT_PUBLIC_LIVEKIT_URL;
 
   if (!apiKey || !apiSecret || !wsUrl) {
-    return NextResponse.json({ error: 'Server misconfigured' }, { status: 500 });
+    return NextResponse.json({ error: "Server misconfigured" }, { status: 500 });
   }
 
   const at = new AccessToken(apiKey, apiSecret, { identity: username });
